requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiofiles
python-dotenv
asyncpg
psutil
requests
chess